+++ affinidi-messaging-mediator/src/common/mod.rs
@@ -1,3 +1,4 @@
+pub mod acl_checks;
 pub mod config;
 pub mod errors;
 pub mod jwt_auth;