+++ affinidi-messaging-sdk/src/errors.rs
@@ -22,6 +22,8 @@ pub enum ATMError {
     ConfigError(String),
     #[error("Authentication error: {0}")]
     AuthenticationError(String),
+    #[error("ACL Denied error: {0}")]
+    ACLDenied(String),
     #[error("DIDComm message error: {0}. Reason: {1}")]
     DidcommError(String, String),
     #[error("SDK Error: {0}")]